Finish splitting up the KtlintIdea and the KtlintBase plugin
