+++ plugin/src/main/kotlin/org/jlleitschuh/gradle/ktlint/KtlintPlugin.kt
@@ -15,11 +15,9 @@ import org.gradle.api.file.SourceDirectorySet
 import org.gradle.api.internal.HasConvention
 import org.gradle.api.plugins.AppliedPlugin
 import org.gradle.api.plugins.Convention
-import org.gradle.api.plugins.HelpTasksPlugin
 import org.gradle.api.plugins.JavaPluginConvention
 import org.gradle.api.tasks.JavaExec
 import org.gradle.api.tasks.StopExecutionException
-import org.gradle.language.base.plugins.LifecycleBasePlugin
 import org.jetbrains.kotlin.gradle.plugin.KonanArtifactContainer
 import org.jetbrains.kotlin.gradle.plugin.KonanExtension
 import org.jetbrains.kotlin.gradle.plugin.KotlinSourceSet
@@ -28,32 +26,17 @@ import org.jlleitschuh.gradle.ktlint.reporter.ReporterType
 import java.io.File
 import kotlin.reflect.KClass
 
-const val VERIFICATION_GROUP = LifecycleBasePlugin.VERIFICATION_GROUP
-const val FORMATTING_GROUP = "Formatting"
-const val HELP_GROUP = HelpTasksPlugin.HELP_GROUP
-const val CHECK_PARENT_TASK_NAME = "ktlintCheck"
-const val FORMAT_PARENT_TASK_NAME = "ktlintFormat"
-const val APPLY_TO_IDEA_TASK_NAME = "ktlintApplyToIdea"
-const val APPLY_TO_IDEA_GLOBALLY_TASK_NAME = "ktlintApplyToIdeaGlobally"
-val KOTLIN_EXTENSIONS = listOf("kt", "kts")
-
 /**
  * Plugin that provides a wrapper over the `ktlint` project.
  */
 open class KtlintPlugin : Plugin<Project> {
 
     override fun apply(target: Project) {
-        target.plugins.apply(KtlintHelperPlugin::class.java)
-
-        // The extension has been added by the helper plugin above.
-        val extension = target.extensions.getByName("ktlint") as KtlintExtension
+        val extension = target.plugins.apply(KtlintBasePlugin::class.java).extension
+        // Apply the idea plugin
+        target.plugins.apply(KtlintIdeaPlugin::class.java)
 
         addKtLintTasksToKotlinPlugin(target, extension)
-
-        // Checking subprojects as well
-        target.subprojects {
-            addKtLintTasksToKotlinPlugin(it, extension)
-        }
     }
 
     private fun addKtLintTasksToKotlinPlugin(target: Project, extension: KtlintExtension) {
@@ -277,76 +260,3 @@ open class KtlintPlugin : Plugin<Project> {
 
     private inline fun <reified T> Convention.getPluginHelper() = getPlugin(T::class.java)
 }
-
-/**
- * Helper plugin that only applies tasks that don't modify the "check" task.
- */
-open class KtlintHelperPlugin : Plugin<Project> {
-    override fun apply(target: Project) {
-        val extension = target.extensions.create("ktlint", KtlintExtension::class.java)
-
-        if (target == target.rootProject) {
-            /*
-             * Only add these tasks if we are applying to the root project.
-             */
-            addApplyToIdeaTasks(target, extension)
-        }
-    }
-
-    private fun addApplyToIdeaTasks(rootProject: Project, extension: KtlintExtension) {
-        rootProject.afterEvaluate {
-            if (rootProject.tasks.findByName(APPLY_TO_IDEA_TASK_NAME) == null) {
-                val ktLintConfig = createConfiguration(rootProject, extension)
-
-                if (extension.isApplyToIdeaPerProjectAvailable()) {
-                    rootProject.taskHelper<KtlintApplyToIdeaTask>(APPLY_TO_IDEA_TASK_NAME) {
-                        group = HELP_GROUP
-                        description = "Generates IDEA built-in formatter rules and apply them to the project." +
-                            "It will overwrite existing ones."
-                        classpath.setFrom(ktLintConfig)
-                        android.set(rootProject.provider { extension.isAndroidFlagEnabled() })
-                        globally.set(rootProject.provider { false })
-                    }
-                }
-
-                rootProject.taskHelper<KtlintApplyToIdeaTask>(APPLY_TO_IDEA_GLOBALLY_TASK_NAME) {
-                    group = HELP_GROUP
-                    description = "Generates IDEA built-in formatter rules and apply them globally " +
-                        "(in IDEA user settings folder). It will overwrite existing ones."
-                    classpath.setFrom(ktLintConfig)
-                    android.set(rootProject.provider { extension.isAndroidFlagEnabled() })
-                    globally.set(rootProject.provider { true })
-                }
-            }
-        }
-    }
-
-    /**
-     * Checks if apply code style to IDEA IDE per project is availalbe.
-     *
-     * Available since KtLint version `0.22.0`.
-     */
-    private fun KtlintExtension.isApplyToIdeaPerProjectAvailable() = SemVer.parse(version) >= SemVer(0, 22, 0)
-}
-
-private fun createConfiguration(target: Project, extension: KtlintExtension) =
-    target.configurations.maybeCreate("ktlint").apply {
-        target.dependencies.add(
-            this.name,
-            mapOf(
-                "group" to "com.github.shyiko",
-                "name" to "ktlint",
-                "version" to extension.version
-            )
-        )
-    }
-
-private inline fun <reified T : Task> Project.taskHelper(name: String, noinline configuration: T.() -> Unit): T {
-    return this.tasks.create(name, T::class.java, configuration)
-}
-
-/**
- * Android option is available from ktlint 0.12.0.
- */
-private fun KtlintExtension.isAndroidFlagEnabled() =
-    android && SemVer.parse(version) >= SemVer(0, 12, 0)
